lifted up states to Dashboard and refactored AuthRoute
